requirements.txt
# ═══════════════════════════════════════════════════════════════
# Agentic Sales Data Analyst & Visualizer — Python Dependencies
# ═══════════════════════════════════════════════════════════════
# Install with:  pip install -r requirements.txt
# Recommended:   Use a virtual environment (venv) before installing.
# ═══════════════════════════════════════════════════════════════

# ---------- LangChain core + integrations ----------
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0        # ChatGoogleGenerativeAI (Gemini)

# ---------- LLM / Embeddings ----------
google-generativeai>=0.8.0           # Google Gemini Python SDK
sentence-transformers>=3.0.0         # Local HuggingFace embeddings (free)

# ---------- Vector Store ----------
chromadb>=0.5.0                      # ChromaDB for RAG vector storage

# ---------- Database ----------
sqlalchemy>=2.0.0                    # SQLAlchemy ORM / engine

# ---------- Data Manipulation ----------
pandas>=2.1.0                        # DataFrame operations

# ---------- Visualization ----------
plotly>=5.18.0                       # Interactive charts (primary)
matplotlib>=3.8.0                    # Static charts (fallback)
kaleido==0.2.1                       # Plotly static PNG export on Windows

# ---------- Fake Data Generation ----------
faker>=28.0.0                        # Realistic synthetic data

# ---------- Environment / Config ----------
python-dotenv>=1.0.0                 # Load .env files
streamlit>=1.30.0                    # Frontend UI
